add llm tokens into different handlers
